Types and Functions

Graphs.SimpleGraphs.SimpleDiGraph — Method

SimpleDiGraph(rf::KirchoffForest)

Convert a KirchoffForest rf to a SimpleDiGraph.

Example

g = grid([3,3])
rf = random_forest(g,.4)
f = SimpleDiGraph(rf)
connected_components(f)

KirchoffForests.RFGraphPlot — Type

mutable struct RFGraphPlot
    g :: AbstractGraph
    xloc :: Array{Number}
    yloc :: Array{Number}
    signal :: Array{Number}
    nodeSize :: Number
    edgeWidth :: Number
    arrowheadwidth :: Number
    cmap :: Symbol
    colorbar :: Bool
    colorbarlabel :: String 
    colorbar_titlefontsize :: Number
    colorbar_tickfontsize :: Number
    title :: String
end

A simple struct for a plotting parameters of graphs and spanning trees/forests

Base.:* — Method

*(rf::KirchoffForest, Y::Matrix)

Treating the random forest as a linear operator, propagate the value of y at the root to the rest of the tree.

Example

g = grid([5])
rf = random_forest(g, .5)
rf*collect(1:nv(g))

Base.:* — Method

*(p::Partition, Y::Matrix)

Treating the graph partition as a linear operator, compute the average of Y over the partition.

Example

g = grid([5])
rf = random_forest(g, .5)
p = Partition(rf)
p*collect(1:nv(g))

KirchoffForests.admm_edge_lasso — Method

admm_edge_lasso(g::AbstractGraph,q::Real,y::AbstractArray;maxiter=100,ρ=0.001,method="exact",Nfor=20)

Implements exact and forest updates of the ADMM algorithm for solving the problem of edge LASSO. See the notebook for an example.

Arguments

g: Input graph

q: Regularization parameter

y: Input signal

Optional Parameters

maxiter: Maximum number of iterations for ADMM

α: Step size for ADMM updates.

method: Method to compute the updates.

method="exact": Computes the update directly.

method="xtilde": Computes the update by using $\tilde{x}$

method="xbar": Computes the update by using $\bar{x}$

Nfor: Number of spanning forests to sample

Outputs

xk: Solution computed by Newton's iteration

loss_func: An array that contains the value of the loss function through the iterations.

KirchoffForests.alias_draw — Method

Drawing procedure of alias method after the preprocessing Its cost is constant!

KirchoffForests.alias_preprocess — Method

This is the implementation of the preprocessing for the alias method.
Overall cost for a graph is O(N^2)
https://en.wikipedia.org/wiki/Alias_method

KirchoffForests.avg_rf — Method

TODO: sum_by function is commented thus not in scope...

KirchoffForests.newton_poisson_noise — Method

newton_poisson_noise(g::AbstractGraph,y::AbstractVector,t0::AbstractVector,mu::Number;α=0.1,numofiter = 100,tol=0.001, method="exact",nrep=100,status=true,line_search=true)

Implements exact and forest updates for solving the problem of Poisson denoising via Laplacian regularization by Newton's method. See the notebook for an illustration.

Arguments

g: Input Graph

y: Input signal

t0: Initial solution

mu: Regularization parameter

Optional Parameters

α: Step size for Newton's update

tol: Tolerance parameter for the increment. Algorithm terminates if $||\mathbf{t}_{k+1} -\mathbf{t}_k||_2<$tol$$

method: Method to compute the updates.

method="exact": Computes the update directly.

method="xtilde": Computes the update by using $\tilde{x}$

method="xbar": Computes the update by using $\bar{x}$

numofiter: Maximum number of iterations for Newton's method

nrep: Number of spanning forests to sample

status: Verbose parameter.

line_search: If true, implements approximate line search for the step size α.

Outputs

sol: Solution computed by Newton's iteration

inc_arr: An array that contains the increment $||\mathbf{t}_{k+1} -\mathbf{t}_k||_2$ through the iterations.

loss_arr: An array that contains the value of loss function through the iterations.

KirchoffForests.next — Method

next(rf::KirchoffForest)

Return a vector of indices v, where v[i] = j means that node i points to node j in the forest. If v[i] = 0 i is a root.

KirchoffForests.partition_boundary_track — Method

partition_boundary_track(g::AbstractGraph,rf::KirchoffForest)

Helper function for computing control variates in trace estimation. Given a forest rf in graph g, this method returns a $n\times 1$ vector boundary_track where boundary_track[i] is the number of neighbors of node $i$ that do not have the same root with $i$ in rf.

KirchoffForests.random_forest — Method

random_forest(G::AbstractGraph,q)

Run Wilson's algorithm on G to generate a random forest with parameter "q". q determines the probability that the random walk is interrupted at a node. If q is a scalar, that probability equals q/(q+d[i]) at node i with degree d[i]. If q is a vector, it equals q[i]/(q[i]+d[i]).

Example

using Graphs G = grid([3,3]) random_forest(G,.4) q_varying = rand(nv(G)) rf = random_forest(G,q_varying) nroots(rf) next(rf) #who points to whom in the forest`

TODO:

Add a rng parameter for reproducibility

Code duplication for random_forest(G::AbstractGraph, q::AbstractFloat/AbstractVector), maybe define a function computing the acceptance to sink node

KirchoffForests.random_spanning_tree — Method

random_spanning_tree(g, [r])

Generate a (uniform) random spanning tree (https://en.wikipedia.org/wiki/Spanning_tree) using Wilson's algorithm (https://dl.acm.org/doi/10.1145/237814.237880). A spanning tree of connected graph g is a connected subgraph of g that's cycle-free, i.e. a tree that includes only edges from g and connects every node. If you specify the root of the tree, the function produces a spanning tree that is picked uniformly among all trees rooted at r. If you do not specify a root, the function produces a random tree from g, picking uniformly among all spanning trees of g (over all possible roots).

NB: a graph must be connected in order to have a spanning tree. By default, the function checks that g is connected (in order to avoid an infinite loop). If you are positive g is connected, use force=true.

Arguments

g: a graph

optional: r, index of a node to serve as root

force: if true, skip connectivity test

Output

If the root is specified, returns a tree, represented as a SimpleDiGraph. If it isn't, returns a named tuple with "tree": the tree and "root": the root.

Examples

julia> g = cycle_graph(4)
{4, 4} undirected simple Int64 graph

julia> random_spanning_tree(g).tree |> edges |> collect
3-element Array{Graphs.SimpleGraphs.SimpleEdge{Int64},1}:
 Edge 2 => 3
 Edge 3 => 4
 Edge 4 => 1

TODO: Maybe consider the function randomspanningtree as a meta function with a switch wilson / aldous..

KirchoffForests.root_boundary_track — Method

root_boundary_track(g::AbstractGraph,rf::KirchoffForest)

Helper function for computing control variates in trace estimation. Given a forest rf in graph g, this method returns a vector boundary_track (in size of the roots of rf) where boundary_track[i] is the number of neighbors of $i$-th root that are not rooted in $i$-th root in rf.

KirchoffForests.smooth — Method

smooth(g::AbstractGraph{T}, q, Y )

Smooth signal over graph. Given a vector $\mathbf{y}$ of size nv(g), compute $q(q\mathbf{I}+\mathbf{L})^{-1}\mathbf{y}$, where $\mathbf{L}$ is the graph Laplacian and q > 0 is a regularisation coefficient (the smaller q, the stronger the smoothing).

If Y is a matrix then this function computes $q(q\mathbf{I}+\mathbf{L})^{-1}\mathbf{Y}$. The linear system is solved using a direct method.

Example

g = grid([10])
t = LinRange(0, 1, 10)
y = sin.(6*pi*t)
smooth(g, .1, y)
smooth(g, 10.1, y)

KirchoffForests.smooth_rf_xbar — Method

smooth_rf_xbar(g :: AbstractGraph,q :: Number,M::SparseMatrixCSC{Float64,Int64},Y::Vector{Float64};maxiter=Inf,abstol=0.0,reltol=0.0,cov_est=false,verbose=false)
smooth_rf_xbar(g :: AbstractGraph,q :: Array{Number},M::SparseMatrixCSC{Float64,Int64},Y::Vector{Float64};maxiter=Inf,abstol=0.0,reltol=0.0,verbose=false)

Memory/time efficient implementations of the RSF-based estimator $\bar{\mathsf{x}}$ for GTR problem.

Arguments

g: Input graph

q: Forest parameter $q$. If it is a scalar, it is the same for every vertex in g.

M: Sparse matrix which equals to $(\mathsf{L}/q +\mathsf{I})$ or $(\mathsf{Q}^{-1}\mathsf{L}+\mathsf{I})$. Only needed for the residual computation.

Y: Input graph signal. One can pass a matrix in size of $n\times k$ instead of a vector to smooth $k$ graph signals.

Optional parameters:

maxiter: Number of forests to sample

abstol: Tolerance for the residual $||\mathsf{M}*\mathsf{Y} - \bar{\mathbf{x}}||_2$. When the residual is less than the tolerance, the algorithm stops.

reltol: Stopping threshold for the relative residual $||\mathsf{M}*\mathsf{Y} - \bar{\mathbf{x}}||_2 / ||\mathsf{Y}||_2$. When the relative residual is less than the tolerance, the algorithm stops.

verbose: Boolean for details. When true, residual error per iteration is reported.

Outputs

est: The estimate given by $\bar{\mathsf{x}}$

indr: Number of iterations taken

Stopping Criterion

When both abstol and reltol are specified, the one that yields a larger tolerance is selected as a stopping criterion.

KirchoffForests.smooth_rf_xtilde — Method

smooth_rf_xtilde(g :: AbstractGraph,q :: Number,M::SparseMatrixCSC{Float64,Int64},Y::Vector{Float64};maxiter=Inf,abstol=0.0,reltol=0.0,cov_est=false,verbose=false)
smooth_rf_xtilde(g :: AbstractGraph,q :: Array{Number},M::SparseMatrixCSC{Float64,Int64},Y::Vector{Float64};maxiter=Inf,abstol=0.0,reltol=0.0,verbose=false)

Memory/time efficient implementations of the RSF-based estimator $\tilde{\mathsf{x}}$ for GTR problem.

Arguments

g: Input graph

q: Forest parameter $q$. If it is a scalar, it is the same for every vertex in g.

M: Sparse matrix which equals to $(\mathsf{L}/q +\mathsf{I})$ or $(\mathsf{Q}^{-1}\mathsf{L}+\mathsf{I})$. Only needed for the residual computation.

Y: Input graph signal. One can pass a matrix in size of $n\times k$ instead of a vector to smooth $k$ graph signals.

Optional parameters:

maxiter: Number of forests to sample

abstol: Tolerance for the residual $||\mathsf{M}*\mathsf{Y} - \tilde{\mathbf{x}}||_2$. When the residual is less than the tolerance, the algorithm stops.

reltol: Stopping threshold for the relative residual $||\mathsf{M}*\mathsf{Y} - \tilde{\mathbf{x}}||_2 / ||\mathsf{Y}||_2$. When the relative residual is less than the tolerance, the algorithm stops.

verbose: Boolean for details. When true, residual error per iteration is reported.

Outputs

est: The estimate given by $\tilde{\mathsf{x}}$

indr: Number of iterations taken

Stopping Criterion

When both abstol and reltol are specified, the one that yields a larger tolerance is selected as a stopping criteria.

KirchoffForests.trace_estimator — Method

trace_estimator(g::AbstractGraph,q::Real;variant=1,α=2*q/(2*q+maximum(degree(g))),NREP=100)

Function to implement the algorithms to estimate $tr(\mathsf{K})$. By modifying the parameter variant, one can reach different algorithms.

Arguments

g: Input graph

q: Forest parameter $q$ (Only implemented for scalar $q$ for now).

Optional parameters

variant: Changes the used forest estimator between four variants.

variant=1: Returns the estimate from number of roots (default).

variant=2: Returns the estimate by $\tilde{s}$.

variant=3: Returns the estimate by $\bar{s}$.

variant=4: Returns the estimate by ${s}_{st}$.

α: The update parameter for $\tilde{s}$ and $\bar{s}$.

NREP: Number of Monte Carlo trials.