Types and Functions

Graphs.SimpleGraphs.SimpleDiGraph — Method

SimpleDiGraph(rf::KirchoffForest)

Convert a KirchoffForest rf to a SimpleDiGraph.

Example

g = grid([3,3])
rf = random_forest(g,.4)
f = SimpleDiGraph(rf)
connected_components(f)

Base.:* — Method

*(rf::KirchoffForest, Y::Matrix)

Treating the random forest as a linear operator, propagate the value of y at the root to the rest of the tree.

Example

g = grid([5])
rf = random_forest(g, .5)
rf*collect(1:nv(g))

Base.:* — Method

*(p::Partition, Y::Matrix)

Treating the graph partition as a linear operator, compute the average of Y over the partition.

Example

g = grid([5])
rf = random_forest(g, .5)
p = Partition(rf)
p*collect(1:nv(g))

KirchoffForests.alias_draw — Method

Drawing procedure of alias method after the preprocessing Its cost is constant!

KirchoffForests.alias_preprocess — Method

This is the implementation of the preprocessing for the alias method.
Overall cost for a graph is O(N^2)
https://en.wikipedia.org/wiki/Alias_method

KirchoffForests.avg_rf — Method

TODO: sum_by function is commented thus not in scope...

KirchoffForests.next — Method

next(rf::KirchoffForest)

Return a vector of indices v, where v[i] = j means that node i points to node j in the forest. If v[i] = 0 i is a root.

KirchoffForests.random_forest — Method

random_forest(G::AbstractGraph,q)

Run Wilson's algorithm on G to generate a random forest with parameter "q". q determines the probability that the random walk is interrupted at a node. If q is a scalar, that probability equals q/(q+d[i]) at node i with degree d[i]. If q is a vector, it equals q[i]/(q[i]+d[i]).

Example

using Graphs G = grid([3,3]) random_forest(G,.4) q_varying = rand(nv(G)) rf = random_forest(G,q_varying) nroots(rf) next(rf) #who points to whom in the forest`

TODO:

Add a rng parameter for reproducibility

Code duplication for random_forest(G::AbstractGraph, q::AbstractFloat/AbstractVector), maybe define a function computing the acceptance to sink node

KirchoffForests.random_spanning_tree — Method

random_spanning_tree(g, [r])

Generate a (uniform) random spanning tree (https://en.wikipedia.org/wiki/Spanning_tree) using Wilson's algorithm (https://dl.acm.org/doi/10.1145/237814.237880). A spanning tree of connected graph g is a connected subgraph of g that's cycle-free, i.e. a tree that includes only edges from g and connects every node. If you specify the root of the tree, the function produces a spanning tree that is picked uniformly among all trees rooted at r. If you do not specify a root, the function produces a random tree from g, picking uniformly among all spanning trees of g (over all possible roots).

NB: a graph must be connected in order to have a spanning tree. By default, the function checks that g is connected (in order to avoid an infinite loop). If you are positive g is connected, use force=true.

Arguments

g: a graph

optional: r, index of a node to serve as root

force: if true, skip connectivity test

Output

If the root is specified, returns a tree, represented as a SimpleDiGraph. If it isn't, returns a named tuple with "tree": the tree and "root": the root.

Examples

julia> g = cycle_graph(4)
{4, 4} undirected simple Int64 graph

julia> random_spanning_tree(g).tree |> edges |> collect
3-element Array{Graphs.SimpleGraphs.SimpleEdge{Int64},1}:
 Edge 2 => 3
 Edge 3 => 4
 Edge 4 => 1

TODO: Maybe consider the function randomspanningtree as a meta function with a switch wilson / aldous..

KirchoffForests.smooth — Method

smooth(g::AbstractGraph{T}, q, Y )

Smooth signal over graph. Given a vector $\mathbf{y}$ of size nv(g), compute $q(q\mathbf{I}+\mathbf{L})^{-1}\mathbf{y}$, where $\mathbf{L}$ is the graph Laplacian and q > 0 is a regularisation coefficient (the smaller q, the stronger the smoothing).

If Y is a matrix then this function computes $q(q\mathbf{I}+\mathbf{L})^{-1}\mathbf{Y}$. The linear system is solved using a direct method.

Example

g = grid([10])
t = LinRange(0, 1, 10)
y = sin.(6*pi*t)
smooth(g, .1, y)
smooth(g, 10.1, y)