Types and Functions

Graphs.SimpleGraphs.SimpleDiGraph — Method

SimpleDiGraph(rf :: RandomForest)

Convert a RandomForest rf to a SimpleDiGraph.

Example

julia> using Graphs, RandomForests

julia> g = grid([3,3])
{9, 12} undirected simple Int64 graph

julia> rf = random_forest(g,.4)
Random forest. Size of original graph 9.
Number of trees 4

julia> f = SimpleDiGraph(rf)
{9, 5} directed simple Int64 graph

julia> connected_components(f)
4-element Array{Array{Int64,1},1}:
 [1, 4]
 [2, 3, 5, 6]
 [7, 8]
 [9]

RandomForests.Partition — Type

p = Partition(rf :: RandomForest )

A simple struct for partition created by random forests. Its fields are:

part: A vector of indices where part[i] returns the partition/tree number that node i belongs to.

nparts: Number of different parts/trees of the partition.

sizep: A vector of trees/parts where sizep[k] returns the size of the k-th tree.

Local Averaging Operator

Local averaging operator within each tree is implemented by using this struct with multiplication. See the multiplication operator * for Partition.

RandomForests.PreprocessedWeightedGraph — Type

g = PreprocessedWeightedGraph(g::SimpleWeightedGraph)
g = PreprocessedWeightedGraph(adjmx::SparseMatrixCSC{U,T})

A special type for weighted graphs. It inherently implements the "alias" method for efficiently sampling discrete random variables in sampling random walks. After a preprocessing ($\mathcal{O}(n^2)$), sampling a random neighbor in a weighted graph becomes $\mathcal{O}(1)$. See Alias method for more details.

RandomForests.RandomForest — Type

rf = RandomForest(next::Array{Int},
                  roots :: BitSet,
                  nroots :: Int,
                  root :: Array{Int,1})

A simple struct for a rooted spanning forest for given arguments.

Arguments

next: a vector of vertices where next[i] = j means that node $i$ points to node $j$ in the forest. If next[i] = 0, $i$ is a root.

roots: the set of roots in the forest.

nroots: the number of the roots.

root: a vector of vertices where root[i] = j means that node $i$ is rooted in node $j$ in the forest.

Base.:* — Method

*(p::Partition,Y :: Matrix)

Treating the graph partition as a linear operator, compute the average of Y over the partition.

Example

g = grid([5])
rf = random_forest(g,.5)
p = Partition(rf)
p*collect(1:nv(g))

Base.:* — Method

*(rf::RandomForest,Y :: Matrix)

Treating the random forest as a linear operator, propagate the value of y at the root to the rest of the tree.

Example

g = grid([5])
rf = random_forest(g,.5)
rf*collect(1:nv(g))

RandomForests.admm_edge_lasso — Method

admm_edge_lasso(g::AbstractGraph,q::Real,y::AbstractArray;maxiter=100,ρ=0.001,method="exact",Nfor=20)

Implements exact and forest updates of the ADMM algorithm for solving the problem of edge LASSO. See the notebook for an example.

Arguments

g: Input graph

q: Regularization parameter

y: Input signal

Optional Parameters

maxiter: Maximum number of iterations for ADMM

α: Step size for ADMM updates.

method: Method to compute the updates.

method="exact": Computes the update directly.

method="xtilde": Computes the update by using $\tilde{x}$

method="xbar": Computes the update by using $\bar{x}$

Nfor: Number of spanning forests to sample

Outputs

xk: Solution computed by Newton's iteration

loss_func: An array that contains the value of the loss function through the iterations.

RandomForests.newton_poisson_noise — Method

newton_poisson_noise(g::AbstractGraph,y::AbstractVector,t0::AbstractVector,mu::Number;α=0.1,numofiter = 100,tol=0.001, method="exact",nrep=100,status=true,line_search=true)

Implements exact and forest updates for solving the problem of Poisson denoising via Laplacian regularization by Newton's method. See the notebook for an illustration.

Arguments

g: Input Graph

y: Input signal

t0: Initial solution

mu: Regularization parameter

Optional Parameters

α: Step size for Newton's update

tol: Tolerance parameter for the increment. Algorithm terminates if $||\mathbf{t}_{k+1} -\mathbf{t}_k||_2<$tol$$

method: Method to compute the updates.

method="exact": Computes the update directly.

method="xtilde": Computes the update by using $\tilde{x}$

method="xbar": Computes the update by using $\bar{x}$

numofiter: Maximum number of iterations for Newton's method

nrep: Number of spanning forests to sample

status: Verbose parameter.

line_search: If true, implements approximate line search for the step size α.

Outputs

sol: Solution computed by Newton's iteration

inc_arr: An array that contains the increment $||\mathbf{t}_{k+1} -\mathbf{t}_k||_2$ through the iterations.

loss_arr: An array that contains the value of loss function through the iterations.

RandomForests.next — Method

v = next(rf::RandomForest)

Return a vector of indices v, where v[i] = j means that node $i$ points to node $j$ in the forest. If v[i] = 0, i is a root.

RandomForests.partition_boundary_track — Method

partition_boundary_track(g::AbstractGraph,rf::RandomForest)

Helper function for computing control variates in trace estimation. Given a forest rf in graph g, this method returns a $n\times 1$ vector boundary_track where boundary_track[i] is the number of neighbors of node $i$ that do not have the same root with $i$ in rf.

RandomForests.random_forest — Method

random_forest(G::AbstractGraph,q[,B::AbstractVector])

Run Wilson's algorithm on G to generate a random forest with parameter $q$.

Arguments

g: Input graph

q: Parameter that determines the probability that the random walk is interrupted at a node. If q is a scalar, that probability equals q/(q+d[i]) at node i with degree d[i]. If q is a vector, it equals q[i]/(q[i]+d[i]).

B: the set of predefined rootset. Random walks are interrupted whenever they reach a node in B.

Example

using Graphs,RandomForests

julia>  G = grid([3,3])
{9, 12} undirected simple Int64 graph

julia> random_forest(G,.4) # Scalar q
Random forest. Size of original graph 9.
Number of trees 4

julia> random_forest(G,rand(nv(G))) # q varies over vertices
Random forest. Size of original graph 9.
Number of trees 2

julia> B = rand(1:nv(G),3) # A predefined rootset.
3-element Array{Int64,1}:
 3
 7
 8

julia> rf = random_forest(G,rand(nv(G)),B)
Random forest. Size of original graph 9.
Number of trees 4

julia> next(rf)
9-element Array{Int64,1}:
 4
 1
 0
 7
 2
 3
 0
 0
 0

julia> rf.roots
BitSet with 4 elements:
 3
 7
 8
 9

RandomForests.random_spanning_tree — Method

random_spanning_tree(g,[r])

Generate a (uniform) random spanning tree using Wilson's algorithm. A spanning tree of g is a connected subgraph of g that's cycle-free, i.e. a tree that includes only edges from g and connects every node. If you specify the root of the tree, the function produces a spanning tree that is picked uniformly among all trees rooted at r. If you do not specify a root, the function produces a random tree from g, picking uniformly among all spanning trees of g (over all possible roots).

NB: a graph must be connected in order to have a spanning tree. By default, the function checks that g is connected (in order to avoid an infinite loop). If you are positive g is connected, use force=true.

Arguments

g: a graph

optional: r, index of a node to serve as the root

force: if true, skip the connectivity test

Output

If the root is specified, returns a tree, represented as a SimpleDiGraph. If it isn't, returns a named tuple with "tree": the tree and "root": the root.

Examples

julia> g = cycle_graph(4)
{4, 4} undirected simple Int64 graph

julia> random_spanning_tree(g).tree |> edges |> collect
3-element Array{Graphs.SimpleGraphs.SimpleEdge{Int64},1}:
 Edge 2 => 3
 Edge 3 => 4
 Edge 4 => 1

RandomForests.random_successor — Method

random_successor(G::AbstractGraph,i :: T) where T <: Int64

Return a neighbor j of node i in the graph g with probability $\frac{w(i,j)}{d_i+q}$. If the graph is preprocessed (see PreprocessedWeightedGraph), the cost is constant.

RandomForests.root_boundary_track — Method

root_boundary_track(g::AbstractGraph,rf::RandomForest)

Helper function for computing control variates in trace estimation. Given a forest rf in graph g, this method returns a vector boundary_track (in size of the roots of rf) where boundary_track[i] is the number of neighbors of $i$-th root that are not rooted in $i$-th root in rf.

RandomForests.smooth — Method

smooth(g :: AbstractGraph{T},q,Y )

Smooth signal over graph. Given a vector $\mathbf{y}$ of size nv(g), compute $q(q\mathsf{I}+\mathsf{L})^{-1}\mathbf{y}$, where $\mathsf{L}$ is the graph Laplacian and q > 0 is a regularisation coefficient (the smaller q, the stronger the smoothing).

If Y is a matrix then this function computes $q(q\mathsf{I}+\mathsf{L})^{-1}\mathsf{Y}$. The linear system is solved using a direct method.

Example

g = grid([10])
t = LinRange(0,1,10)
y = sin.(6*pi*t)
smooth(g,.1,y)
smooth(g,10.1,y)

RandomForests.smooth_rf — Function

smooth_rf(g :: AbstractGraph,q :: Float64,Y,rootset=[];nrep=10,variant=1,mean_correction=false)
smooth_rf(g :: AbstractGraph,q :: Vector,Y,rootset=[];nrep=10,variant=1,mean_correction=false)

This function implements RSF-based estimators for GTR (i.e. $\mathsf{K}\mathsf{Y}$) and trace of $\mathsf{K}$.

Arguments

g: Input graph.

q: Regularization parameter. It is either a scalar or an $n$-dimensional vector. In the latter, regularization parameter is defined node-wise.

Y: Input graph signal. Can be an $n$-dimensional vector. If it is a matrix, the estimators estimate $\mathsf{K}\mathsf{Y}$.

rootset: Predefined root set (Default is empty set.). The roots of the sampled forests must include rootset. In terms of GTR, this setting corresponds to the limit where the regularization parameters $q_i$s goes to infinity for the nodes in rootset. In other words, no smoothing applied in these nodes.

Optional parameters

nrep: Number of forests to sample.

variant: Parameter to choose the estimator. Set variant=1 for $\tilde{\mathbf{x}}$ and variant=2 for $\bar{\mathbf{x}}$.

Outputs:

est: Estimate for GTR problem.

nroots: Estimate for $tr(\mathsf{K})$.

diagK: Estimate for diagonals of $\mathsf{K}$.

RandomForests.smooth_rf_xbar — Method

smooth_rf_xbar(g :: AbstractGraph,q :: Number,M::SparseMatrixCSC{Float64,Int64},Y::Vector{Float64};maxiter=Inf,abstol=0.0,reltol=0.0,cov_est=false,verbose=false)
smooth_rf_xbar(g :: AbstractGraph,q :: Array{Number},M::SparseMatrixCSC{Float64,Int64},Y::Vector{Float64};maxiter=Inf,abstol=0.0,reltol=0.0,verbose=false)

Memory/time efficient implementations of the RSF-based estimator $\bar{\mathsf{x}}$ for GTR problem.

Arguments

g: Input graph

q: Forest parameter $q$. If it is a scalar, it is the same for every vertex in g.

M: Sparse matrix which equals to $(\mathsf{L}/q +\mathsf{I})$ or $(\mathsf{Q}^{-1}\mathsf{L}+\mathsf{I})$. Only needed for the residual computation.

Y: Input graph signal. One can pass a matrix in size of $n\times k$ instead of a vector to smooth $k$ graph signals.

Optional parameters:

maxiter: Number of forests to sample

abstol: Tolerance for the residual $||\mathsf{M}*\mathsf{Y} - \bar{\mathbf{x}}||_2$. When the residual is less than the tolerance, the algorithm stops.

reltol: Stopping threshold for the relative residual $||\mathsf{M}*\mathsf{Y} - \bar{\mathbf{x}}||_2 / ||\mathsf{Y}||_2$. When the relative residual is less than the tolerance, the algorithm stops.

verbose: Boolean for details. When true, residual error per iteration is reported.

Outputs

est: The estimate given by $\bar{\mathsf{x}}$

indr: Number of iterations taken

Stopping Criterion

When both abstol and reltol are specified, the one that yields a larger tolerance is selected as a stopping criterion.

RandomForests.smooth_rf_xtilde — Method

smooth_rf_xtilde(g :: AbstractGraph,q :: Number,M::SparseMatrixCSC{Float64,Int64},Y::Vector{Float64};maxiter=Inf,abstol=0.0,reltol=0.0,cov_est=false,verbose=false)
smooth_rf_xtilde(g :: AbstractGraph,q :: Array{Number},M::SparseMatrixCSC{Float64,Int64},Y::Vector{Float64};maxiter=Inf,abstol=0.0,reltol=0.0,verbose=false)

Memory/time efficient implementations of the RSF-based estimator $\tilde{\mathsf{x}}$ for GTR problem.

Arguments

g: Input graph

q: Forest parameter $q$. If it is a scalar, it is the same for every vertex in g.

M: Sparse matrix which equals to $(\mathsf{L}/q +\mathsf{I})$ or $(\mathsf{Q}^{-1}\mathsf{L}+\mathsf{I})$. Only needed for the residual computation.

Y: Input graph signal. One can pass a matrix in size of $n\times k$ instead of a vector to smooth $k$ graph signals.

Optional parameters:

maxiter: Number of forests to sample

abstol: Tolerance for the residual $||\mathsf{M}*\mathsf{Y} - \tilde{\mathbf{x}}||_2$. When the residual is less than the tolerance, the algorithm stops.

reltol: Stopping threshold for the relative residual $||\mathsf{M}*\mathsf{Y} - \tilde{\mathbf{x}}||_2 / ||\mathsf{Y}||_2$. When the relative residual is less than the tolerance, the algorithm stops.

verbose: Boolean for details. When true, residual error per iteration is reported.

Outputs

est: The estimate given by $\tilde{\mathsf{x}}$

indr: Number of iterations taken

Stopping Criterion

When both abstol and reltol are specified, the one that yields a larger tolerance is selected as a stopping criteria.

RandomForests.trace_estimator — Method

trace_estimator(g::AbstractGraph,q::Real;variant=1,α=2*q/(2*q+maximum(degree(g))),NREP=100)

Function to implement the algorithms to estimate $tr(\mathsf{K})$. By modifying the parameter variant, one can reach different algorithms.

Arguments

g: Input graph

q: Forest parameter $q$ (Only implemented for scalar $q$ for now).

Optional parameters

variant: Changes the used forest estimator between four variants.

variant=1: Returns the estimate from number of roots (default).

variant=2: Returns the estimate by $\tilde{s}$.

variant=3: Returns the estimate by $\bar{s}$.

variant=4: Returns the estimate by ${s}_{st}$.

α: The update parameter for $\tilde{s}$ and $\bar{s}$.

NREP: Number of Monte Carlo trials.